requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.1
mdurl==0.1.2
multidict==6.1.0
narwhals==1.28.0
numpy==2.2.3
openai==1.82.0
openai-agents==0.0.16
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
streamlit==1.42.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.2
uvicorn==0.34.2
watchdog==6.0.0
yarl==1.18.3
